requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
channels>=4.0
channels-redis>=4.1
daphne>=4.0
celery>=5.3
redis>=5.0
django-cors-headers>=4.3
Pillow>=10.0
sentence-transformers>=2.2
openai>=1.6
